add open tracing to pulsar client
